requirements.txt
boto3==1.4.6
botocore==1.6.1
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cycler==0.10.0
descartes==1.1.0
docutils==0.14
Fiona==1.7.8
geopandas==0.2.1
jmespath==0.9.3
matplotlib==2.0.2
munch==2.2.0
numpy==1.13.1
pandas==0.20.3
pyparsing==2.2.0
pyproj==1.9.5.1
python-dateutil==2.6.1
python-dotenv==0.6.5
pytz==2017.2
s3transfer==0.1.10
Shapely==1.5.17.post1
six==1.10.0
